+++ src/loaders/NDS2SFLoader.cpp
@@ -43,7 +43,7 @@ PostLoadCommand NDS2SFLoader::Apply(RawFile* file)
 			//pRoot->UI_WriteBufferToFile(L"uncomp.nds", exebuf, exebufsize);
 
 			wstring str = file->GetFileName();
-			pRoot->CreateVirtFile(exebuf, exebufsize, str.data());
+			pRoot->CreateVirtFile(exebuf, (uint32_t)exebufsize, str.data());
 			return DELETE_IT;
 		}
 	}
